refactor(lyrics): separate the dom builder from the injection lifecycle
